feat: mvvm userinfo split header and verification
